+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -1,12 +1,17 @@
-20201115:
+20201116:
 3: 
 4: 
 5: 
 
+20201115:
+3: 0983, 1272
+4: 
+5: 
+
 20201114:
-3: 0307, 0731
+3: 0253, 0307, 0729, 0731
 4: 0732
-5: 0315
+5: 0315, 1649
 
 20201113:
 3: 0105, 0106, 0889, 1109
