+++ src/main/java/com/zea7ot/leetcode/ood/lvl3/lc0731/SolutionApproach0SegmentedTree.java
@@ -1,5 +1,5 @@
 /**
- * https://leetcode.com/problems/my-calendar-i/
+ * https://leetcode.com/problems/my-calendar-ii/
  * 
  * Time Complexity:     O(N * lg(`RANGE`))
  * Space Complexity:    O(N * lg(`RANGE`))
@@ -7,30 +7,34 @@
  * References:
  *  https://leetcode.com/problems/my-calendar-ii/discuss/109528/nlogd-Java-solution-using-segment-tree-with-lazy-propagation-(for-the-general-case-of-K-booking)
  */
-package com.zea7ot.leetcode.lvl3.lc0729;
+package com.zea7ot.leetcode.ood.lvl3.lc0731;
 
 public class SolutionApproach0SegmentedTree {
-    private SegmentedTreeNode root;
+    private static final int RANGE = (int)1e9;
+
+    private SegmentTreeNode root;
 
     public SolutionApproach0SegmentedTree() {
-        final int RANGE = (int) 1e9;
-        this.root = new SegmentedTreeNode(0, RANGE, 0);
+        this.root = new SegmentTreeNode(0, RANGE, 0);
     }
 
     public boolean book(int start, int end) {
+        // sanity check
         if (start > end)
             return false;
 
         int k = query(root, start, end - 1);
-        // to return false if any interval occupied
-        if (k > 0)
+        // to generalize:
+        // to replace '2' with 'K - 1',
+        // 'K' is the actual allowed overlaps
+        if (k >= 2)
             return false;
 
         update(root, start, end - 1, 1);
         return true;
     }
 
-    private int query(SegmentedTreeNode node, final int RANGE_LOW, final int RANGE_HIGH) {
+    private int query(SegmentTreeNode node, final int RANGE_LOW, final int RANGE_HIGH) {
         if (RANGE_LOW > RANGE_HIGH || node == null || RANGE_LOW > node.HIGH || RANGE_HIGH < node.LOW)
             return 0;
 
@@ -42,7 +46,7 @@ public class SolutionApproach0SegmentedTree {
         return Math.max(query(node.left, RANGE_LOW, RANGE_HIGH), query(node.right, RANGE_LOW, RANGE_HIGH));
     }
 
-    private void update(SegmentedTreeNode node, final int RANGE_LOW, final int RANGE_HIGH, int val) {
+    private void update(SegmentTreeNode node, final int RANGE_LOW, final int RANGE_HIGH, int val) {
         if (RANGE_LOW > RANGE_HIGH || node == null || RANGE_LOW > node.HIGH || RANGE_HIGH < node.LOW)
             return;
 
@@ -60,16 +64,16 @@ public class SolutionApproach0SegmentedTree {
         node.k = Math.max(node.left.k, node.right.k);
     }
 
-    private void normalize(SegmentedTreeNode node) {
+    private void normalize(SegmentTreeNode node) {
         if (node.LOW < node.HIGH) {
             if (node.left == null || node.right == null) {
                 final int MID = node.LOW + (node.HIGH - node.LOW) / 2;
 
-                node.left = new SegmentedTreeNode(node.LOW, MID, node.k);
-                node.right = new SegmentedTreeNode(MID + 1, node.HIGH, node.k);
+                node.left = new SegmentTreeNode(node.LOW, MID, node.k);
+                node.right = new SegmentTreeNode(MID + 1, node.HIGH, node.k);
             } else if (node.lazy > 0) {
                 node.left.k += node.lazy;
-                node.right.lazy = node.lazy;
+                node.left.lazy += node.lazy;
 
                 node.right.k += node.lazy;
                 node.right.lazy += node.lazy;
@@ -79,15 +83,16 @@ public class SolutionApproach0SegmentedTree {
         node.lazy = 0;
     }
 
-    private class SegmentedTreeNode {
+    private class SegmentTreeNode {
         private final int LOW, HIGH;
         private int k, lazy;
-        private SegmentedTreeNode left, right;
+        private SegmentTreeNode left, right;
 
-        private SegmentedTreeNode(final int LOW, final int HIGH, int k) {
+        private SegmentTreeNode(final int LOW, final int HIGH, int k) {
             this.LOW = LOW;
             this.HIGH = HIGH;
             this.k = k;
+            this.lazy = 0;
             this.left = this.right = null;
         }
     }
