+++ src/main/java/com/zea7ot/leetcode/tags/by_algorithm/range_query/segment_tree.txt
@@ -17,6 +17,7 @@ References:
 4   0732    https://leetcode.com/problems/my-calendar-iii/
 
 5   0315    https://leetcode.com/problems/count-of-smaller-numbers-after-self/
+5   1649    https://leetcode.com/problems/create-sorted-array-through-instructions/
