+++ src/main/java/com/zea7ot/leetcode/tags/by_algorithm/range_query/summary.txt
@@ -23,5 +23,6 @@ References:
 4   1292    https://leetcode.com/problems/maximum-side-length-of-a-square-with-sum-less-than-or-equal-to-threshold/
 
 5   0315    https://leetcode.com/problems/count-of-smaller-numbers-after-self/
-
+5   1649    https://leetcode.com/problems/create-sorted-array-through-instructions/
+s
 6   0308    https://leetcode.com/problems/range-sum-query-2d-mutable/
