+++ src/main/java/com/zea7ot/leetcode/tags/by_algorithm/sort/summary.txt
@@ -52,6 +52,7 @@ References:
 5   0363    https://leetcode.com/problems/max-sum-of-rectangle-no-larger-than-k/
 5   0774    https://leetcode.com/problems/minimize-max-distance-to-gas-station/
 5   0786    https://leetcode.com/problems/k-th-smallest-prime-fraction/
+5   1649    https://leetcode.com/problems/create-sorted-array-through-instructions/
