+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/tree/by_data_structure/binary_search_tree.txt
@@ -21,6 +21,7 @@ References:
 4   1373    https://leetcode.com/problems/maximum-sum-bst-in-binary-tree/
 
 5   0315    https://leetcode.com/problems/count-of-smaller-numbers-after-self/
+5   1649    https://leetcode.com/problems/create-sorted-array-through-instructions/
 
 
 Resource:
