+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/tree/by_data_structure/segment_tree.txt
@@ -18,7 +18,7 @@ References:
 
 5   0315    https://leetcode.com/problems/count-of-smaller-numbers-after-self/
 5   0715    https://leetcode.com/problems/range-module/
-
+5   1649    https://leetcode.com/problems/create-sorted-array-through-instructions/
 
 
 ### Resources
