+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/tree/summary.txt
@@ -152,3 +152,4 @@ References:
 5   0315    https://leetcode.com/problems/count-of-smaller-numbers-after-self/
 5   0715    https://leetcode.com/problems/range-module/
 5   0834    https://leetcode.com/problems/sum-of-distances-in-tree/
+5   1649    https://leetcode.com/problems/create-sorted-array-through-instructions/
